INV-0-1-0 : finalizing v0.1.0, refactor, add documentation README.md, basic examples, unit test

+++ tests/inventory_test.rs
@@ -3,7 +3,7 @@ mod tests {
     use inventory_kit::error::InventoryError;
     use inventory_kit::in_memory;
     use inventory_kit::model::InventoryItem;
-    use inventory_kit::service::{AtomicInventoryOps, InventoryRepository};
+    use inventory_kit::repository::{AtomicInventoryOps, InventoryRepository};
 
     #[derive(Clone, Debug, PartialEq, Eq, Hash)]
     struct HotelRoom {
@@ -20,7 +20,7 @@ mod tests {
 
     #[test]
     fn test_reserve_and_release() {
-        let mut repo = in_memory::InMemoryInventoryRepository::<HotelRoom, u64>::new();
+        let mut repo = in_memory::InMemoryInventoryRepository::<HotelRoom, u32>::new();
 
         let room = HotelRoom { id: 1 };
         repo.insert_availability(room.id(), 1000, 2000, 10);
@@ -45,26 +45,23 @@ mod tests {
 
     #[test]
     fn test_atomic_reserve_success_and_fail() {
-        let mut repo = in_memory::InMemoryInventoryRepository::<HotelRoom, u64>::new();
+        let mut repo = in_memory::InMemoryInventoryRepository::<HotelRoom, u32>::new();
         let room_id = 1;
         repo.insert_availability(room_id, 1000, 1100, 5);
         repo.insert_availability(room_id, 1100, 1200, 3);
 
-        // Should succeed
         let result = repo.reserve_many(&room_id, &[
             (1000, 1100, 2),
             (1100, 1200, 1),
         ]);
         assert!(result.is_ok());
 
-        // Should fail (not enough in second slot)
         let result = repo.reserve_many(&room_id, &[
             (1000, 1100, 2),
-            (1100, 1200, 5), // not enough
+            (1100, 1200, 5),
         ]);
         assert_eq!(result, Err(InventoryError::Insufficient));
 
-        // Ensure nothing changed after failed reserve
         let slots = repo.get_availability(&room_id, 1000, 1200).unwrap();
         assert_eq!(slots[0].available, 3); // 5 - 2
         assert_eq!(slots[1].available, 2); // 3 - 1
